WIP Perform long-running initialization tasks asynchronously.
TODO: Finishing writing tests.

Fixes #43.

When adbfs-automount launches adbfs to mount a newly-started device, the
device may not have mounted the device root path yet. With async init,
adbfs will wait up to a minute for the target to be mounted before terminating
with an error.

This is a slightly bigger change because it requires changing how each of the
file operation methods starts: each method must now wait for the initialization
to finish, and check for any errors it reported.